[#75] chore: useViewportresize hook 분리
